Organize source files
