+++ SphericalWorldGenerator/DataTypes/River.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace SphericalWorldGenerator
+namespace SphericalWorldGenerator.DataTypes
 {
     public enum Direction
     {
